eliminar/ml-pipeline/requirements.txt
ultralytics==8.0.234
opencv-python==4.8.1.78
numpy==1.24.3
PyYAML==6.0
Pillow==10.1.0
roboflow==1.1.22
